Update AW_ApsEditor and AW_Aps classes.
A new text label is added to the AW_ApsEditor setup content to clarify its purpose. In AW_Aps, the component menu and class names are updated to reflect that it's no longer solely hole-focused but functions as an 'APS Manager'. The help URL in AW_ApsRingMarker has been removed.